+++ src/it/resources/postgres/inventory.sql
@@ -1,3 +1,2 @@
--- Create the schema that we'll use to populate data and watch the effect in the binlog
 CREATE SCHEMA inventory;
-SET search_path TO inventory;
+ALTER DATABASE postgres SET search_path = inventory;