+++ src/it/scala/ai/tripl/arc/transform/MySQLDebeziumTransformSuite.scala
@@ -7,6 +7,7 @@ import java.util.UUID
 
 import scala.collection.JavaConverters._
 import scala.util.Random
+import scala.util.control.Breaks._
 
 import org.scalatest.FunSuite
 import org.scalatest.BeforeAndAfter
@@ -22,6 +23,7 @@ import ai.tripl.arc.api._
 import ai.tripl.arc.api.API._
 import ai.tripl.arc.util.log.LoggerFactory
 import ai.tripl.arc.udf.UDF
+import ai.tripl.arc.transform.DebeziumStringKafkaEvent
 
 import ai.tripl.arc.util._
 
@@ -31,10 +33,13 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
 
   val inputView = "inputView"
   val outputView = "outputView"
+  val initialStateView = "initialStateView"
   val schema = "schema"
   val checkpointLocation = "/tmp/debezium"
+  val serverName = "dbserver1"
+  val size = 5000
 
-  val mysqlURL = "jdbc:mysql://mysql:3306/inventory?user=root&password=debezium&allowMultiQueries=true"
+  val databaseURL = "jdbc:mysql://mysql:3306/inventory?user=root&password=debezium&allowMultiQueries=true"
   val connectURI = s"http://connect:8083/connectors/"
   val connectorName = "inventory-connector-mysql"
 
@@ -44,7 +49,9 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
                   .master("local[*]")
                   .config("spark.ui.port", "4040")
                   .config("spark.checkpoint.compress", "true")
+                  .config("spark.sql.shuffle.partitions", 4)
                   .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
+                  .config("spark.kryoserializer.buffer.max", "2047m")
                   .config("spark.sql.streaming.checkpointLocation", checkpointLocation)
                   .config("spark.sql.streaming.forceDeleteTempCheckpointLocation", "true")
                   .appName("Arc Test")
@@ -67,7 +74,7 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         println("Query terminated: " + queryTerminated.id)
       }
       override def onQueryProgress(queryProgress: QueryProgressEvent): Unit = {
-        println(s"numRowsTotal: ${queryProgress.progress.stateOperators(0).numRowsTotal} inputRowsPerSecond: ${queryProgress.progress.inputRowsPerSecond.round} processedRowsPerSecond: ${queryProgress.progress.processedRowsPerSecond.round}")
+        println(s"numRowsTotal: ${if (queryProgress.progress.stateOperators.length == 0) 0 else queryProgress.progress.stateOperators(0).numRowsTotal} inputRowsPerSecond: ${queryProgress.progress.inputRowsPerSecond.round} processedRowsPerSecond: ${queryProgress.progress.processedRowsPerSecond.round}")
       }
     })
 
@@ -90,7 +97,7 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
     |    "database.user": "debezium",
     |    "database.password": "dbz",
     |    "database.server.id": "184054",
-    |    "database.server.name": "dbserver2",
+    |    "database.server.name": "${serverName}",
     |    "database.whitelist": "inventory",
     |    "database.history.kafka.bootstrap.servers": "kafka:9092",
     |    "database.history.kafka.topic": "schema-changes.inventory",
@@ -101,7 +108,8 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
   }
 
   def makeTransaction(statements: Seq[String]): String = {
-    s"""START TRANSACTION;
+    s"""SET TRANSACTION ISOLATION LEVEL READ COMMITTED;
+    |START TRANSACTION;
     |${statements.mkString("\n")}
     |COMMIT;
     """.stripMargin
@@ -185,13 +193,13 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
     (transactions, updates, inserts, deletes)
   }
 
-  test("MySQLDebeziumTransform") {
+  test("MySQLDebeziumTransform: Streaming") {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = TestUtils.getLogger()
     implicit val arcContext = TestUtils.getARCContext(isStreaming = true)
 
-    val (customerInitial, customerUpdates) = TestHelpers.getTestData(50000)
+    val (customerInitial, customerUpdates) = TestHelpers.getTestData(size)
     val customersMetadata = MetadataUtils.createMetadataDataframe(customerInitial.toDF)
     customersMetadata.persist
     customersMetadata.createOrReplaceTempView(schema)
@@ -208,14 +216,14 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             id=None,
             name="JDBCExecute",
             description=None,
-            inputURI=new URI(mysqlURL),
-            jdbcURL=mysqlURL,
+            inputURI=new URI(databaseURL),
+            jdbcURL=databaseURL,
             sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (c_custkey INTEGER PRIMARY KEY NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);")),
             params=Map.empty,
             sqlParams=Map.empty
           )
         )
-        customerInitial.write.mode("append").jdbc(mysqlURL, s"inventory.${tableName}", new java.util.Properties)
+        customerInitial.write.mode("append").jdbc(databaseURL, s"inventory.${tableName}", new java.util.Properties)
 
         // make transactions
         val (transactions, update, insert, delete) = makeTransactions(customerInitial, customerUpdates, tableName, seed)
@@ -226,7 +234,7 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
           .readStream
           .format("kafka")
           .option("kafka.bootstrap.servers", "kafka:9092")
-          .option("subscribe", s"dbserver2.inventory.${tableName}")
+          .option("subscribe", s"${serverName}.inventory.${tableName}")
           .option("startingOffsets", "earliest")
           .load
         readStream.createOrReplaceTempView(inputView)
@@ -240,7 +248,12 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             inputView=inputView,
             outputView=outputView,
             schema=Left(schema),
-            strict=strict
+            strict=strict,
+            initialStateView=None,
+            initialStateKey=None,
+            persist=false,
+            numPartitions=None,
+            partitionBy=List.empty,
           )
         )
 
@@ -255,45 +268,58 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
           // wait for query to start
           val start = System.currentTimeMillis()
           while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
-            if (System.currentTimeMillis() > start + 180000) throw new Exception("Timeout without messages arriving")
+            if (System.currentTimeMillis() > start + 60000) throw new Exception("Timeout without messages arriving")
             println("Waiting for query progress...")
             Thread.sleep(1000)
           }
 
-          // while running perform PARALLEL insert/update/delete transactions
+          // while running perform SERIAL insert/update/delete transactions
           // this will block the main thread but we want to process all updates before triggering awaitTermination
           var last = System.currentTimeMillis()
           var i = 0
           var deadlocks = 0
-          transactions.par.foreach { sql =>
+          transactions.foreach { sql =>
             if (System.currentTimeMillis() > last+1000) {
               last = System.currentTimeMillis()
-              println(s"${i} transactions/sec")
+              println(s"${i} transactions/sec (${deadlocks} deadlocks)")
               i = 0
             }
             i += 1
-            try {
-              ai.tripl.arc.execute.JDBCExecuteStage.execute(
-                ai.tripl.arc.execute.JDBCExecuteStage(
-                  plugin=new ai.tripl.arc.execute.JDBCExecute,
-                  id=None,
-                  name="JDBCExecute",
-                  description=None,
-                  inputURI=new URI(mysqlURL),
-                  jdbcURL=mysqlURL,
-                  sql=sql,
-                  params=Map.empty,
-                  sqlParams=Map.empty
-                )
-              )
-            } catch {
-              // not nice but sometimes get deadlocks and we can just ignore them
-              case e: Exception if e.getMessage.contains("Deadlock") => deadlocks += 1
+            var retry = 0
+            breakable {
+              while(true){
+                if (retry == 100) {
+                  throw new Exception("could not complete transaciton due to deadlocks")
+                  break
+                }
+                try {
+                  ai.tripl.arc.execute.JDBCExecuteStage.execute(
+                    ai.tripl.arc.execute.JDBCExecuteStage(
+                      plugin=new ai.tripl.arc.execute.JDBCExecute,
+                      id=None,
+                      name="JDBCExecute",
+                      description=None,
+                      inputURI=new URI(databaseURL),
+                      jdbcURL=databaseURL,
+                      sql=sql,
+                      params=Map.empty,
+                      sqlParams=Map.empty
+                    )
+                  )
+                  break
+                } catch {
+                  case e: Exception if e.getMessage.contains("Deadlock found") => {
+                    retry += 1
+                    deadlocks += 1
+                    Thread.sleep(200)
+                  }
+                }
+              }
             }
           }
+          println(s"executed ${transactions.length} transactions (${deadlocks} deadlocks) against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes.")
 
-          if (deadlocks > 10) throw new Exception(s"too many (${deadlocks}) transactions ignored due to database deadlock")
-
+          Thread.sleep(5000)
           writeStream.processAllAvailable
           writeStream.stop
 
@@ -306,8 +332,8 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
               description=None,
               schema=Right(Nil),
               outputView="expected",
-              jdbcURL=mysqlURL,
-              driver=DriverManager.getDriver(mysqlURL),
+              jdbcURL=databaseURL,
+              driver=DriverManager.getDriver(databaseURL),
               tableName=s"inventory.${tableName}",
               numPartitions=None,
               fetchsize=None,
@@ -320,8 +346,8 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             )
           ).get
           assert(TestUtils.datasetEquality(expected, spark.table(tableName)))
+          println("PASS\n")
 
-          println(s"executed ${transactions.length} transactions against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes\n")
         } catch {
           case e: Exception => fail(e.getMessage)
         } finally {
@@ -333,8 +359,8 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
               id=None,
               name="JDBCExecute",
               description=None,
-              inputURI=new URI(mysqlURL),
-              jdbcURL=mysqlURL,
+              inputURI=new URI(databaseURL),
+              jdbcURL=databaseURL,
               sql=makeTransaction(Seq(s"DROP TABLE inventory.${tableName};")),
               params=Map.empty,
               sqlParams=Map.empty
@@ -365,15 +391,15 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         id=None,
         name="JDBCExecute",
         description=None,
-        inputURI=new URI(mysqlURL),
-        jdbcURL=mysqlURL,
+        inputURI=new URI(databaseURL),
+        jdbcURL=databaseURL,
         sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (booleanDatum BOOLEAN NOT NULL, dateDatum DATE NOT NULL, decimalDatum DECIMAL(10,3) NOT NULL, doubleDatum DOUBLE NOT NULL, integerDatum INTEGER NOT NULL, longDatum BIGINT NOT NULL, stringDatum VARCHAR(255) NOT NULL, timeDatum VARCHAR(255) NOT NULL, timestampDatum TIMESTAMP NOT NULL);")),
         params=Map.empty,
         sqlParams=Map.empty
       )
     )
 
-    knownData.write.mode("append").jdbc(mysqlURL, s"inventory.${tableName}", new java.util.Properties)
+    knownData.write.mode("append").jdbc(databaseURL, s"inventory.${tableName}", new java.util.Properties)
 
     TestHelpers.registerConnector(connectURI, makeConnectorConfig(s"inventory.${tableName}", "integerDatum"))
 
@@ -381,7 +407,7 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
       .readStream
       .format("kafka")
       .option("kafka.bootstrap.servers", "kafka:9092")
-      .option("subscribe", s"dbserver2.inventory.${tableName}")
+      .option("subscribe", s"${serverName}.inventory.${tableName}")
       .option("startingOffsets", "earliest")
       .load
     readStream.createOrReplaceTempView(inputView)
@@ -395,7 +421,12 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         inputView=inputView,
         outputView=outputView,
         schema=Left(schema),
-        strict=true
+        strict=true,
+        initialStateView=None,
+        initialStateKey=None,
+        persist=false,
+        numPartitions=None,
+        partitionBy=List.empty,
       )
     )
 
@@ -410,11 +441,12 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
       // wait for query to start
       val start = System.currentTimeMillis()
       while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
-        if (System.currentTimeMillis() > start + 180000) throw new Exception("Timeout without messages arriving")
+        if (System.currentTimeMillis() > start + 60000) throw new Exception("Timeout without messages arriving")
         println("Waiting for query progress...")
         Thread.sleep(1000)
       }
 
+      Thread.sleep(5000)
       writeStream.processAllAvailable
       writeStream.stop
 
@@ -431,14 +463,232 @@ class MySQLDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
           id=None,
           name="JDBCExecute",
           description=None,
-          inputURI=new URI(mysqlURL),
-          jdbcURL=mysqlURL,
+          inputURI=new URI(databaseURL),
+          jdbcURL=databaseURL,
           sql=makeTransaction(Seq(s"DROP TABLE inventory.${tableName};")),
           params=Map.empty,
           sqlParams=Map.empty
         )
       )
       writeStream.stop
+    }
+  }
+
+  test("MySQLDebeziumTransform: Batch") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = TestUtils.getLogger()
+    implicit val arcContext = TestUtils.getARCContext(isStreaming = true)
+
+    val (customerInitial, customerUpdates) = TestHelpers.getTestData(size)
+    val customersMetadata = MetadataUtils.createMetadataDataframe(customerInitial.toDF)
+    customersMetadata.persist
+    customersMetadata.createOrReplaceTempView(schema)
+
+    println()
+    for (seed <- 0 to 0) {
+      for (strict <- Seq(true, false)) {
+        val tableName = s"customers_${UUID.randomUUID.toString.replaceAll("-","")}"
+        println(s"mysql ${if (strict) "strict" else "not-strict"} seed: ${seed} target: ${tableName}")
+
+        ai.tripl.arc.execute.JDBCExecuteStage.execute(
+          ai.tripl.arc.execute.JDBCExecuteStage(
+            plugin=new ai.tripl.arc.execute.JDBCExecute,
+            id=None,
+            name="JDBCExecute",
+            description=None,
+            inputURI=new URI(databaseURL),
+            jdbcURL=databaseURL,
+            sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (c_custkey INTEGER NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);")),
+            params=Map.empty,
+            sqlParams=Map.empty
+          )
+        )
+        customerInitial.write.mode("append").jdbc(databaseURL, tableName, new java.util.Properties)
+        ai.tripl.arc.execute.JDBCExecuteStage.execute(
+          ai.tripl.arc.execute.JDBCExecuteStage(
+            plugin=new ai.tripl.arc.execute.JDBCExecute,
+            id=None,
+            name="JDBCExecute",
+            description=None,
+            inputURI=new URI(databaseURL),
+            jdbcURL=databaseURL,
+            sql=makeTransaction(Seq(s"ALTER TABLE ${tableName} ADD PRIMARY KEY (c_custkey);")),
+            params=Map.empty,
+            sqlParams=Map.empty
+          )
+        )
+
+        // make transactions
+        val (transactions, update, insert, delete) = makeTransactions(customerInitial, customerUpdates, tableName, seed)
+
+        TestHelpers.registerConnector(connectURI, makeConnectorConfig(s"inventory.${tableName}", "c_custkey"))
+
+        val readStream = spark
+          .readStream
+          .format("kafka")
+          .option("kafka.bootstrap.servers", "kafka:9092")
+          .option("subscribe", s"${serverName}.inventory.${tableName}")
+          .option("startingOffsets", "earliest")
+          .load
+        readStream.createOrReplaceTempView(inputView)
+
+        val writeStream = readStream
+          .writeStream
+          .outputMode("append")
+          .queryName(tableName)
+          .format("memory")
+          .start
+
+        try {
+          // wait for query to start
+          val start = System.currentTimeMillis()
+          while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
+            if (System.currentTimeMillis() > start + 60000) throw new Exception("Timeout without messages arriving")
+            println("Waiting for query progress...")
+            Thread.sleep(1000)
+          }
+
+          // while running perform SERIAL insert/update/delete transactions
+          // this will block the main thread but we want to process all updates before triggering awaitTermination
+          var last = System.currentTimeMillis()
+          var i = 0
+          var deadlocks = 0
+          transactions.foreach { sql =>
+            if (System.currentTimeMillis() > last+1000) {
+              last = System.currentTimeMillis()
+              println(s"${i} transactions/sec (${deadlocks} deadlocks)")
+              i = 0
+            }
+            i += 1
+            var retry = 0
+            breakable {
+              while(true){
+                if (retry == 100) {
+                  throw new Exception("could not complete transaciton due to deadlocks")
+                  break
+                }
+                try {
+                  ai.tripl.arc.execute.JDBCExecuteStage.execute(
+                    ai.tripl.arc.execute.JDBCExecuteStage(
+                      plugin=new ai.tripl.arc.execute.JDBCExecute,
+                      id=None,
+                      name="JDBCExecute",
+                      description=None,
+                      inputURI=new URI(databaseURL),
+                      jdbcURL=databaseURL,
+                      sql=sql,
+                      params=Map.empty,
+                      sqlParams=Map.empty
+                    )
+                  )
+                  break
+                } catch {
+                  case e: Exception if e.getMessage.contains("Deadlock found") => {
+                    retry += 1
+                    deadlocks += 1
+                    Thread.sleep(200)
+                  }
+                }
+              }
+            }
+          }
+          println(s"executed ${transactions.length} transactions (${deadlocks} deadlocks) against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes.")
+
+          Thread.sleep(5000)
+          writeStream.processAllAvailable
+          writeStream.stop
+
+          // read in batch mode
+          val read = spark
+            .read
+            .format("kafka")
+            .option("kafka.bootstrap.servers", "kafka:9092")
+            .option("subscribe", s"${serverName}.inventory.${tableName}")
+            .option("startingOffsets", "earliest")
+            .load
+            .as[DebeziumStringKafkaEvent]
+            .collect
+            .sortBy(row => row.offset)
+            .toList
+
+          // make sure the updates have happened
+          assert(read.size > customerInitial.count)
+
+          // recursively apply the records passing in the previous state
+          val batches = 3
+          println(s"processing ${read.size} events in ${batches} ${read.size/batches} record batches...")
+          read.grouped(read.size/batches).zipWithIndex.foreach { case (batch, index) =>
+            val start = System.currentTimeMillis()
+            batch.toDF.createOrReplaceTempView(inputView)
+            transform.DebeziumTransformStage.execute(
+              transform.DebeziumTransformStage(
+                plugin=new transform.DebeziumTransform,
+                id=None,
+                name="DebeziumTransform",
+                description=None,
+                inputView=inputView,
+                outputView=outputView,
+                schema=Left(schema),
+                strict=strict,
+                initialStateView=if (index == 0) None else Option(initialStateView),
+                initialStateKey=Option("c_custkey"),
+                persist=true,
+                numPartitions=None,
+                partitionBy=List.empty,
+              )
+            )
+            val output = spark.table(outputView)
+            output.createOrReplaceTempView(initialStateView)
+            println(s"processed batch ${index} of ${batch.length} records in ${System.currentTimeMillis()-start}ms... new total ${output.count}")
+          }
+
+          // validate results
+          val expected = extract.JDBCExtractStage.execute(
+            extract.JDBCExtractStage(
+              plugin=new extract.JDBCExtract,
+              id=None,
+              name="dataset",
+              description=None,
+              schema=Right(Nil),
+              outputView="expected",
+              jdbcURL=databaseURL,
+              driver=DriverManager.getDriver(databaseURL),
+              tableName=s"inventory.${tableName}",
+              numPartitions=None,
+              fetchsize=None,
+              partitionBy=Nil,
+              customSchema=None,
+              persist=true,
+              partitionColumn=None,
+              predicates=Nil,
+              params=Map.empty
+            )
+          ).get
+          assert(TestUtils.datasetEquality(expected, spark.table(outputView)))
+          println(s"PASS: expected: ${expected.count} actual: ${spark.table(outputView).count}\n")
+
+        } catch {
+          case e: Exception => fail(e.getMessage)
+        } finally {
+          TestHelpers.deleteConnector(connectURI, connectorName)
+          TestHelpers.deleteConnector(connectURI, s"${connectorName}-dbhistory")
+          ai.tripl.arc.execute.JDBCExecuteStage.execute(
+            ai.tripl.arc.execute.JDBCExecuteStage(
+              plugin=new ai.tripl.arc.execute.JDBCExecute,
+              id=None,
+              name="JDBCExecute",
+              description=None,
+              inputURI=new URI(databaseURL),
+              jdbcURL=databaseURL,
+              sql=makeTransaction(Seq(s"DROP TABLE inventory.${tableName};")),
+              params=Map.empty,
+              sqlParams=Map.empty
+            )
+          )
+          writeStream.stop
+        }
+      }
     }
   }