+++ src/it/scala/ai/tripl/arc/transform/PostgresDebeziumTransformSuite.scala
@@ -7,6 +7,7 @@ import java.util.UUID
 
 import scala.collection.JavaConverters._
 import scala.util.Random
+import scala.util.control.Breaks._
 
 import org.scalatest.FunSuite
 import org.scalatest.BeforeAndAfter
@@ -22,6 +23,7 @@ import ai.tripl.arc.api._
 import ai.tripl.arc.api.API._
 import ai.tripl.arc.util.log.LoggerFactory
 import ai.tripl.arc.udf.UDF
+import ai.tripl.arc.transform.DebeziumStringKafkaEvent
 
 import ai.tripl.arc.util._
 
@@ -31,10 +33,13 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
 
   val inputView = "inputView"
   val outputView = "outputView"
+  val initialStateView = "initialStateView"
   val schema = "schema"
   val checkpointLocation = "/tmp/debezium"
+  val serverName = "dbserver1"
+  val size = 5000
 
-  val postgresURL = "jdbc:postgresql://postgres:5432/?user=postgres&password=postgres"
+  val databaseURL = "jdbc:postgresql://postgres:5432/postgres?currentSchema=inventory&user=postgres&password=postgres"
   val connectURI = s"http://connect:8083/connectors/"
   val connectorName = "inventory-connector-postgres"
 
@@ -44,7 +49,9 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
                   .master("local[*]")
                   .config("spark.ui.port", "4040")
                   .config("spark.checkpoint.compress", "true")
+                  .config("spark.sql.shuffle.partitions", 8)
                   .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
+                  .config("spark.kryoserializer.buffer.max", "2047m")
                   .config("spark.sql.streaming.checkpointLocation", checkpointLocation)
                   .config("spark.sql.streaming.forceDeleteTempCheckpointLocation", "true")
                   .appName("Arc Test")
@@ -67,7 +74,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         println("Query terminated: " + queryTerminated.id)
       }
       override def onQueryProgress(queryProgress: QueryProgressEvent): Unit = {
-        println(s"numRowsTotal: ${queryProgress.progress.stateOperators(0).numRowsTotal} inputRowsPerSecond: ${queryProgress.progress.inputRowsPerSecond.round} processedRowsPerSecond: ${queryProgress.progress.processedRowsPerSecond.round}")
+        println(s"numRowsTotal: ${if (queryProgress.progress.stateOperators.length == 0) 0 else queryProgress.progress.stateOperators(0).numRowsTotal} inputRowsPerSecond: ${queryProgress.progress.inputRowsPerSecond.round} processedRowsPerSecond: ${queryProgress.progress.processedRowsPerSecond.round}")
       }
     })
 
@@ -90,18 +97,21 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
     |    "database.user": "postgres",
     |    "database.password": "postgres",
     |    "database.dbname": "postgres",
-    |    "database.server.name": "dbserver1",
+    |    "database.server.name": "${serverName}",
     |    "database.history.kafka.bootstrap.servers": "kafka:9092",
     |    "database.history.kafka.topic": "schema-changes.inventory",
-    |    "schema.whitelist": "inventory",
+    |    "schema.include.list": "inventory",
     |    "message.key.columns": "${tableName}:${key}",
-    |    "decimal.handling.mode": "string"
+    |    "decimal.handling.mode": "string",
+    |    "heartbeat.interval​.ms": 20,
+    |    "slot.drop.on.stop": "true",
+    |    "snapshot.mode": "always"
     |  }
     |}""".stripMargin
   }
 
   def makeTransaction(statements: Seq[String]): String = {
-    s"""START TRANSACTION;
+    s"""BEGIN TRANSACTION ISOLATION LEVEL SERIALIZABLE;
     |${statements.mkString("\n")}
     |COMMIT;
     """.stripMargin
@@ -185,13 +195,13 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
     (transactions, updates, inserts, deletes)
   }
 
-  test("PostgresDebeziumTransform") {
+  test("PostgresDebeziumTransform: Streaming") {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = TestUtils.getLogger()
     implicit val arcContext = TestUtils.getARCContext(isStreaming = true)
 
-    val (customerInitial, customerUpdates) = TestHelpers.getTestData(50000)
+    val (customerInitial, customerUpdates) = TestHelpers.getTestData(size)
     val customersMetadata = MetadataUtils.createMetadataDataframe(customerInitial.toDF)
     customersMetadata.persist
     customersMetadata.createOrReplaceTempView(schema)
@@ -208,14 +218,19 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             id=None,
             name="JDBCExecute",
             description=None,
-            inputURI=new URI(postgresURL),
-            jdbcURL=postgresURL,
-            sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (c_custkey INTEGER PRIMARY KEY NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);")),
+            inputURI=new URI(databaseURL),
+            jdbcURL=databaseURL,
+            sql=makeTransaction(
+              Seq(
+                s"CREATE TABLE ${tableName} (c_custkey INTEGER NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);",
+                s"ALTER TABLE ${tableName} REPLICA IDENTITY FULL;"
+              )
+            ),
             params=Map.empty,
             sqlParams=Map.empty
           )
         )
-        customerInitial.write.mode("append").jdbc(postgresURL, s"inventory.${tableName}", new java.util.Properties)
+        customerInitial.write.mode("append").jdbc(databaseURL, s"${tableName}", new java.util.Properties)
 
         // make transactions
         val (transactions, update, insert, delete) = makeTransactions(customerInitial, customerUpdates, tableName, seed)
@@ -226,7 +241,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
           .readStream
           .format("kafka")
           .option("kafka.bootstrap.servers", "kafka:9092")
-          .option("subscribe", s"dbserver1.inventory.${tableName}")
+          .option("subscribe", s"${serverName}.inventory.${tableName}")
           .option("startingOffsets", "earliest")
           .load
         readStream.createOrReplaceTempView(inputView)
@@ -240,7 +255,12 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
             inputView=inputView,
             outputView=outputView,
             schema=Left(schema),
-            strict=strict
+            strict=strict,
+            initialStateView=None,
+            initialStateKey=None,
+            persist=true,
+            numPartitions=None,
+            partitionBy=List.empty,
           )
         )
 
@@ -255,37 +275,52 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
           // wait for query to start
           val start = System.currentTimeMillis()
           while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
-            if (System.currentTimeMillis() > start + 180000) throw new Exception("Timeout without messages arriving")
+            if (System.currentTimeMillis() > start + 60000) throw new Exception("Timeout without messages arriving")
             println("Waiting for query progress...")
             Thread.sleep(1000)
           }
 
-          // while running perform PARALLEL insert/update/delete transactions
+          // while running perform SERIAL insert/update/delete transactions
           // this will block the main thread but we want to process all updates before triggering awaitTermination
           var last = System.currentTimeMillis()
           var i = 0
-          transactions.par.foreach { sql =>
+          var deadlocks = 0
+          transactions.foreach { sql =>
             if (System.currentTimeMillis() > last+1000) {
               last = System.currentTimeMillis()
-              println(s"${i} transactions/sec")
+              println(s"${i} transactions/sec (${deadlocks} deadlocks)")
               i = 0
             }
             i += 1
-            ai.tripl.arc.execute.JDBCExecuteStage.execute(
-              ai.tripl.arc.execute.JDBCExecuteStage(
-                plugin=new ai.tripl.arc.execute.JDBCExecute,
-                id=None,
-                name="JDBCExecute",
-                description=None,
-                inputURI=new URI(postgresURL),
-                jdbcURL=postgresURL,
-                sql=sql,
-                params=Map.empty,
-                sqlParams=Map.empty
-              )
-            )
+            breakable {
+              while(true){
+                try {
+                  ai.tripl.arc.execute.JDBCExecuteStage.execute(
+                    ai.tripl.arc.execute.JDBCExecuteStage(
+                      plugin=new ai.tripl.arc.execute.JDBCExecute,
+                      id=None,
+                      name="JDBCExecute",
+                      description=None,
+                      inputURI=new URI(databaseURL),
+                      jdbcURL=databaseURL,
+                      sql=sql,
+                      params=Map.empty,
+                      sqlParams=Map.empty
+                    )
+                  )
+                  break
+                } catch {
+                  case e: Exception if e.getMessage.contains("could not serialize access") => {
+                    deadlocks += 1
+                    Thread.sleep(200)
+                  }
+                }
+              }
+            }
           }
+          println(s"executed ${transactions.length} transactions (${deadlocks} deadlocks) against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes.")
 
+          Thread.sleep(5000)
           writeStream.processAllAvailable
           writeStream.stop
 
@@ -298,8 +333,8 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
               description=None,
               schema=Right(Nil),
               outputView="expected",
-              jdbcURL=postgresURL,
-              driver=DriverManager.getDriver(postgresURL),
+              jdbcURL=databaseURL,
+              driver=DriverManager.getDriver(databaseURL),
               tableName=s"inventory.${tableName}",
               numPartitions=None,
               fetchsize=None,
@@ -311,9 +346,10 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
               params=Map.empty
             )
           ).get
+          assert(expected.count > customerInitial.count)
           assert(TestUtils.datasetEquality(expected, spark.table(tableName)))
+          println("PASS\n")
 
-          println(s"executed ${transactions.length} transactions against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes\n")
         } catch {
           case e: Exception => fail(e.getMessage)
         } finally {
@@ -325,8 +361,8 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
               id=None,
               name="JDBCExecute",
               description=None,
-              inputURI=new URI(postgresURL),
-              jdbcURL=postgresURL,
+              inputURI=new URI(databaseURL),
+              jdbcURL=databaseURL,
               sql=makeTransaction(Seq(s"DROP TABLE inventory.${tableName};")),
               params=Map.empty,
               sqlParams=Map.empty
@@ -357,15 +393,20 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         id=None,
         name="JDBCExecute",
         description=None,
-        inputURI=new URI(postgresURL),
-        jdbcURL=postgresURL,
-        sql=makeTransaction(Seq(s"CREATE TABLE ${tableName} (booleanDatum BOOLEAN NOT NULL, dateDatum DATE NOT NULL, decimalDatum DECIMAL(10,3) NOT NULL, doubleDatum DOUBLE PRECISION NOT NULL, integerDatum INTEGER NOT NULL, longDatum BIGINT NOT NULL, stringDatum VARCHAR(255) NOT NULL, timeDatum VARCHAR(255) NOT NULL, timestampDatum TIMESTAMP NOT NULL);")),
+        inputURI=new URI(databaseURL),
+        jdbcURL=databaseURL,
+        sql=makeTransaction(
+          Seq(
+            s"CREATE TABLE ${tableName} (booleanDatum BOOLEAN NOT NULL, dateDatum DATE NOT NULL, decimalDatum DECIMAL(10,3) NOT NULL, doubleDatum DOUBLE PRECISION NOT NULL, integerDatum INTEGER NOT NULL, longDatum BIGINT NOT NULL, stringDatum VARCHAR(255) NOT NULL, timeDatum VARCHAR(255) NOT NULL, timestampDatum TIMESTAMP NOT NULL);",
+            s"ALTER TABLE ${tableName} REPLICA IDENTITY FULL;"
+          )
+        ),
         params=Map.empty,
         sqlParams=Map.empty
       )
     )
 
-    knownData.write.mode("append").jdbc(postgresURL, s"inventory.${tableName}", new java.util.Properties)
+    knownData.write.mode("append").jdbc(databaseURL, s"${tableName}", new java.util.Properties)
 
     TestHelpers.registerConnector(connectURI, makeConnectorConfig(s"inventory.${tableName}", "integerDatum"))
 
@@ -373,7 +414,7 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
       .readStream
       .format("kafka")
       .option("kafka.bootstrap.servers", "kafka:9092")
-      .option("subscribe", s"dbserver1.inventory.${tableName}")
+      .option("subscribe", s"${serverName}.inventory.${tableName}")
       .option("startingOffsets", "earliest")
       .load
     readStream.createOrReplaceTempView(inputView)
@@ -387,7 +428,12 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
         inputView=inputView,
         outputView=outputView,
         schema=Left(schema),
-        strict=true
+        strict=true,
+        initialStateView=None,
+        initialStateKey=None,
+        persist=true,
+        numPartitions=None,
+        partitionBy=List.empty,
       )
     )
 
@@ -402,11 +448,12 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
       // wait for query to start
       val start = System.currentTimeMillis()
       while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
-        if (System.currentTimeMillis() > start + 180000) throw new Exception("Timeout without messages arriving")
+        if (System.currentTimeMillis() > start + 60000) throw new Exception("Timeout without messages arriving")
         println("Waiting for query progress...")
         Thread.sleep(1000)
       }
 
+      Thread.sleep(5000)
       writeStream.processAllAvailable
       writeStream.stop
 
@@ -422,14 +469,231 @@ class PostgresDebeziumTransformSuite extends FunSuite with BeforeAndAfter {
           id=None,
           name="JDBCExecute",
           description=None,
-          inputURI=new URI(postgresURL),
-          jdbcURL=postgresURL,
-          sql=makeTransaction(Seq(s"DROP TABLE inventory.${tableName};")),
+          inputURI=new URI(databaseURL),
+          jdbcURL=databaseURL,
+          sql=makeTransaction(Seq(s"DROP TABLE ${tableName};")),
           params=Map.empty,
           sqlParams=Map.empty
         )
       )
       writeStream.stop
+    }
+  }
+
+  test("PostgresDebeziumTransform: Batch") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = TestUtils.getLogger()
+    implicit val arcContext = TestUtils.getARCContext(isStreaming = true)
+
+    val (customerInitial, customerUpdates) = TestHelpers.getTestData(size)
+    val customersMetadata = MetadataUtils.createMetadataDataframe(customerInitial.toDF)
+    customersMetadata.persist
+    customersMetadata.createOrReplaceTempView(schema)
+
+    println()
+    for (seed <- 0 to 0) {
+      for (strict <- Seq(true, false)) {
+        val tableName = s"customers_${UUID.randomUUID.toString.replaceAll("-","")}"
+        println(s"postgres ${if (strict) "strict" else "not-strict"} seed: ${seed} target: ${tableName}")
+
+        ai.tripl.arc.execute.JDBCExecuteStage.execute(
+          ai.tripl.arc.execute.JDBCExecuteStage(
+            plugin=new ai.tripl.arc.execute.JDBCExecute,
+            id=None,
+            name="JDBCExecute",
+            description=None,
+            inputURI=new URI(databaseURL),
+            jdbcURL=databaseURL,
+            sql=makeTransaction(
+              Seq(
+                s"CREATE TABLE ${tableName} (c_custkey INTEGER NOT NULL, c_name VARCHAR(25) NOT NULL, c_address VARCHAR(40) NOT NULL, c_nationkey INTEGER NOT NULL, c_phone VARCHAR(15) NOT NULL, c_acctbal DECIMAL(20,2) NOT NULL, c_mktsegment VARCHAR(10) NOT NULL, c_comment VARCHAR(117) NOT NULL);",
+                s"ALTER TABLE ${tableName} REPLICA IDENTITY FULL;"
+              )
+            ),
+            params=Map.empty,
+            sqlParams=Map.empty
+          )
+        )
+        customerInitial.write.mode("append").jdbc(databaseURL, tableName, new java.util.Properties)
+        ai.tripl.arc.execute.JDBCExecuteStage.execute(
+          ai.tripl.arc.execute.JDBCExecuteStage(
+            plugin=new ai.tripl.arc.execute.JDBCExecute,
+            id=None,
+            name="JDBCExecute",
+            description=None,
+            inputURI=new URI(databaseURL),
+            jdbcURL=databaseURL,
+            sql=makeTransaction(Seq(s"ALTER TABLE ${tableName} ADD PRIMARY KEY (c_custkey);")),
+            params=Map.empty,
+            sqlParams=Map.empty
+          )
+        )
+
+        // make transactions
+        val (transactions, update, insert, delete) = makeTransactions(customerInitial, customerUpdates, tableName, seed)
+
+        TestHelpers.registerConnector(connectURI, makeConnectorConfig(s"inventory.${tableName}", "c_custkey"))
+
+        val readStream = spark
+          .readStream
+          .format("kafka")
+          .option("kafka.bootstrap.servers", "kafka:9092")
+          .option("subscribe", s"${serverName}.inventory.${tableName}")
+          .option("startingOffsets", "earliest")
+          .load
+        readStream.createOrReplaceTempView(inputView)
+
+        val writeStream = readStream
+          .writeStream
+          .outputMode("append")
+          .queryName(tableName)
+          .format("memory")
+          .start
+
+        try {
+          // wait for query to start
+          val start = System.currentTimeMillis()
+          while (writeStream.lastProgress == null || (writeStream.lastProgress != null && writeStream.lastProgress.numInputRows == 0)) {
+            if (System.currentTimeMillis() > start + 60000) throw new Exception("Timeout without messages arriving")
+            println("Waiting for query progress...")
+            Thread.sleep(1000)
+          }
+
+          // while running perform SERIAL insert/update/delete transactions
+          // this will block the main thread but we want to process all updates before triggering awaitTermination
+          var last = System.currentTimeMillis()
+          var i = 0
+          var deadlocks = 0
+          transactions.foreach { sql =>
+            if (System.currentTimeMillis() > last+1000) {
+              last = System.currentTimeMillis()
+              println(s"${i} transactions/sec (${deadlocks} deadlocks)")
+              i = 0
+            }
+            i += 1
+            breakable {
+              while(true){
+                try {
+                  ai.tripl.arc.execute.JDBCExecuteStage.execute(
+                    ai.tripl.arc.execute.JDBCExecuteStage(
+                      plugin=new ai.tripl.arc.execute.JDBCExecute,
+                      id=None,
+                      name="JDBCExecute",
+                      description=None,
+                      inputURI=new URI(databaseURL),
+                      jdbcURL=databaseURL,
+                      sql=sql,
+                      params=Map.empty,
+                      sqlParams=Map.empty
+                    )
+                  )
+                  break
+                } catch {
+                  case e: Exception if e.getMessage.contains("could not serialize access") => {
+                    deadlocks += 1
+                    Thread.sleep(100)
+                  }
+                }
+              }
+            }
+          }
+          println(s"executed ${transactions.length} transactions (${deadlocks} deadlocks) against ${tableName} with ${update} updates, ${insert} inserts, ${delete} deletes.")
+
+          Thread.sleep(5000)
+          writeStream.processAllAvailable
+          writeStream.stop
+
+          // read in batch mode
+          val read = spark
+            .read
+            .format("kafka")
+            .option("kafka.bootstrap.servers", "kafka:9092")
+            .option("subscribe", s"${serverName}.inventory.${tableName}")
+            .option("startingOffsets", "earliest")
+            .load
+            .as[DebeziumStringKafkaEvent]
+            .collect
+            .sortBy(row => row.offset)
+            .toList
+
+          // make sure the updates have happened
+          assert(read.size > customerInitial.count)
+
+          // recursively apply the records passing in the previous state
+          val batches = 3
+          println(s"processing ${read.size} events in ${batches} ${read.size/batches} record batches...")
+          read.grouped(read.size/batches).zipWithIndex.foreach { case (batch, index) =>
+            val start = System.currentTimeMillis()
+            batch.toDF.createOrReplaceTempView(inputView)
+            transform.DebeziumTransformStage.execute(
+              transform.DebeziumTransformStage(
+                plugin=new transform.DebeziumTransform,
+                id=None,
+                name="DebeziumTransform",
+                description=None,
+                inputView=inputView,
+                outputView=outputView,
+                schema=Left(schema),
+                strict=strict,
+                initialStateView=if (index == 0) None else Option(initialStateView),
+                initialStateKey=Option("c_custkey"),
+                persist=true,
+                numPartitions=None,
+                partitionBy=List.empty,
+              )
+            )
+            val output = spark.table(outputView)
+            output.createOrReplaceTempView(initialStateView)
+            println(s"processed batch ${index} of ${batch.length} records in ${System.currentTimeMillis()-start}ms... new total ${output.count}")
+          }
+
+          // validate results
+          val expected = extract.JDBCExtractStage.execute(
+            extract.JDBCExtractStage(
+              plugin=new extract.JDBCExtract,
+              id=None,
+              name="dataset",
+              description=None,
+              schema=Right(Nil),
+              outputView="expected",
+              jdbcURL=databaseURL,
+              driver=DriverManager.getDriver(databaseURL),
+              tableName=s"inventory.${tableName}",
+              numPartitions=None,
+              fetchsize=None,
+              partitionBy=Nil,
+              customSchema=None,
+              persist=true,
+              partitionColumn=None,
+              predicates=Nil,
+              params=Map.empty
+            )
+          ).get
+          assert(TestUtils.datasetEquality(expected, spark.table(outputView)))
+          println(s"PASS: expected: ${expected.count} actual: ${spark.table(outputView).count}\n")
+
+        } catch {
+          case e: Exception => fail(e.getMessage)
+        } finally {
+          TestHelpers.deleteConnector(connectURI, connectorName)
+          TestHelpers.deleteConnector(connectURI, s"${connectorName}-dbhistory")
+          ai.tripl.arc.execute.JDBCExecuteStage.execute(
+            ai.tripl.arc.execute.JDBCExecuteStage(
+              plugin=new ai.tripl.arc.execute.JDBCExecute,
+              id=None,
+              name="JDBCExecute",
+              description=None,
+              inputURI=new URI(databaseURL),
+              jdbcURL=databaseURL,
+              sql=makeTransaction(Seq(s"DROP TABLE ${tableName};")),
+              params=Map.empty,
+              sqlParams=Map.empty
+            )
+          )
+          writeStream.stop
+        }
+      }
     }
   }