+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.1.0
+
+- remove constraint to only support `streaming` mode.
+- add support for injecting a previous state by providing `initialStateView` and `initialStateKey`. currently only supports single value keys.
+- add support for `persist`, `numPartitions` and `partitionBy` to align with standard Arc extract stages.
+
 ## 1.0.0
 
 - initial release.